update scripts: only execute debug or running when compiling succeed

+++ ADT/C++/gdbcpp.sh
@@ -2,10 +2,13 @@ if test -e ./$1.cpp
 then
 	echo g++ -g -std=c++14 $1.cpp -lm -o $1
 	g++ -g -std=c++14 $1.cpp -lm -o $1
-	echo running gdb
-	gdb $1
-	echo delete output file
-	rm $1
+	if [ $? -eq 0 ]
+	then
+		echo running gdb
+		gdb $1
+		echo delete output file
+		rm $1
+	fi
 else
 	echo Error: illegal source file name.
 	echo Usage: ./gdbcpp.sh file_name_without_.cpp
